macos 10.15 update

+++ MESCHACH_CPGPLOT/SRCS/graphics1D_pgplot_transient_xt_window.c
@@ -42,7 +42,7 @@ typedef struct XAW_WINPGPLOT_
    int pgplot_win;
 
    /* Xt id workproc */
-   int id_idleworkproc;
+   unsigned long id_idleworkproc;
 
    /* pointer to widgets */
    Widget * buttonStopRun;
