+++ MESCHACH_VOGLE/SRCS/graphics1D_voplcontour_transient_xt_window.c
@@ -37,7 +37,7 @@ typedef struct XAW_WINVOGLE_
    int vogle_win;
 	
    /* Xt id workproc */
-   int id_idleworkproc;
+   unsigned long id_idleworkproc;
 	
    /* pointer to widgets */
    Widget * buttonStopRun;
@@ -563,7 +563,7 @@ void * vopl_contours_with_xt_toolkit_transient(void* data)
    /* --------------------------------------------------------------------------------- */
 	
    int vogle_win1 = -1;
-   int id_idleworkproc;
+   unsigned long id_idleworkproc;
    /* on this separate thread, the Xt window ... */
    int ac    = 0;
    char **av = NULL;
